bug fixes and code clean up
fixed a bug where if there was one value left in the list, it would iterate and add a boat X amount of times as the initial input for the amount of fishermen you would insert. The fix was creating an if statement that would check to see if there was only one value left and then break out after adding the final boat to the count.